Removing "entrepreneurs"
Based on past discussions with the group, I think it'd be best to
leave this word out of our description in order to avoid giving
people the wrong ideas about who we are. If you disagree with this
change bring it up at the next meeting, as I think this is an important
change to who we are.